Return None if the underlying C Pointer is NULL
Some LLVM C functions return a NULL pointer (for example when
converting values using the `is_a_...` family of functions. If the
instruction is not of the more concrete type, the C layer returns a
NULL pointer (`== ffi.NULL`). However, this Null pointer got wrapped
in a Python Value object, resulting in a crash for some functions.

The cleaner approach implemented in this patch is to return None (in
`__new__` if the underlying pointer is NULL). Then, any access to
this invalid value is caught by the runtime, resulting in an
`AttributeError`.