Improve handling for sized string return values
Some LLVM functions like `LLVMGetMDString` return a char pointer
and write the length of the returned string into a pointer to
unsigned int/long argument. This case is now handled correctly
as long as the length argument is the function's last argument.
Instead of the length argument, the Python function gets an optional
parameter to set the encoding of the returned string or to directly
return a bytestring.

To ease finding the correct parameters a global list of unsigned
integer types is added.

Testcase: Create a LLVM metadata string from a Python string
containing null bytes and roundtrip back to a Python string using
`value.get_md_string()`. Also check the bytestring result.