Improve the way we look for programs
We introduced a function to look for a program which employs
`shutil.which`. In addition to the standard `PATH`, we also inspect the
LLVM binary directory (i.e., the output of `llvm-config --bin-dir`).

We also prioritize looking for `clang` over the system `cpp` program,
which, at least on macOS, might be problematic.